Refactored Postlist and HomePage
Moved onScroll event listeners to PostList as they pertain to that component
Made PostList a functional component for access to the component lifecycle
moved postList__header and children to an inclusive-functional-component inside the main PostList component. It's a lot of code and not likely to change. I'm trying to clean up the render method.
Debating utilizing a heavier state on this component, possibly separating each post as well.